--- src/modules/components/proposal/proposalActions/proposalActionsTypes/ProposalActionType.ts
@@ -1,4 +1,0 @@
-export enum ProposalActionType {
-    WITHDRAW_TOKEN = 'withdrawToken',
-    UPDATE_METADATA = 'updateMetadata'
-}
