+++ src/modules/components/proposal/proposalActions/proposalActionsTypes/index.ts
@@ -1,5 +1,5 @@
 export type { IMetadataLink, IMetaData, IProposalActionUpdateMetadata } from './proposalActionUpdateMetadata';
 export type { IProposalAction, IProposalActionInputData, IProposalActionInputDataParameter } from './proposalAction';
 export { ProposalActionComponent, type IProposalActionComponentProps } from './proposalActionComponent';
-export { ProposalActionType } from './proposalActionType';
+export { ProposalActionType } from './tempProposalActionType';
 export type { IProposalActionWithdrawToken } from './proposalActionWithdrawToken';
