+++ src/modules/components/proposal/proposalActions/proposalActionsTypes/proposalActionUpdateMetadata.ts
@@ -1,5 +1,5 @@
 import { type IProposalAction } from "./proposalAction";
-import { type ProposalActionType } from "./proposalActionType";
+import { type ProposalActionType } from "./tempProposalActionType";
 
 export interface IMetadataLink {
   /**
