+++ src/modules/components/proposal/proposalActions/proposalActionsTypes/proposalActionWithdrawToken.ts
@@ -1,6 +1,6 @@
 import type { ICompositeAddress } from '../../../../types';
 import type { IProposalAction } from './proposalAction';
-import type { ProposalActionType } from './proposalActionType';
+import type { ProposalActionType } from './tempProposalActionType';
 
 export interface IProposalActionWithdrawToken extends IProposalAction {
     /**
